+++ VocaVocca/View/Voca/VocaMainViewController.swift
@@ -12,7 +12,7 @@ class VocaMainViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         // Do any additional setup after loading the view.
-        view.backgroundColor = .brown
+        view.backgroundColor = .white
     }
